Update link stock to prog
Changed some descriptions of key variables
Linked Stock Report to immediately follow Prog Report. 
There is a problem in that I ask for week number again. but if I remove this, then we cannot request only a stock report. 
All requests would be to redo entire report. This would lead to a lot of missing data.